+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-22T14:03:04.810Z",
+    "2026-06-23T06:18:34.109Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {