+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -118,7 +118,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KVQT9DDSZQ65610R6F56HWRF"),
+  _id: z.string().default("01KVSJ3CN0SZSX84WGTXFPEW08"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: RetrieveDatasourceStatus$inboundSchema,