+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -152,7 +152,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KVQT9DDZBSB2AGKHAGF5T495"),
+  _id: z.string().default("01KVSJ3CN5GVNFTNG59G8V127V"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,