+++ packages/orq-rc/RELEASES.md
@@ -10189,3 +10189,13 @@ Based on:
 - [typescript v4.11.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.40 - packages/orq-rc
+
+## 2026-06-23 06:21:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.784.1 (2.911.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.41 - packages/orq-rc